refactor(tests): split the three modules over the line limit
test_llm_providers, test_tool_runtime_contract and test_run_tests_tool
had grown to 1003/897/841 lines. Split each along the seams its own
section banners already marked -- provider classification vs request
shaping vs response harvesting; runtime and bash vs file tools vs
search; output parsing vs runner selection vs green-evidence integrity
-- and lift the doubles each pair now shares into a *_test_support
module, matching the convention the suite already uses.

Pure relocation: every collected test node survives under the same name,
1738 before and 1738 after excluding the new hygiene cases.
