+++ tests/test_tool_runtime_contract.py
@@ -1,15 +1,11 @@
-import asyncio
-import subprocess
-import time
 from pathlib import Path
-from types import SimpleNamespace
 
 import pytest
 from tool_execution_test_support import AlwaysAllowPermissionPolicy as FakePermissionPolicy
+from tool_runtime_test_support import FakeEnv, SpySafetyPolicy, run
 
-from opencollab.adapters.env import LocalEnvironment
 from opencollab.adapters.safety import SandboxInterceptor
-from opencollab.adapters.tools import base, human
+from opencollab.adapters.tools import human
 from opencollab.adapters.tools.apply_patch import ApplyPatchTool
 from opencollab.adapters.tools.base import Tool
 from opencollab.adapters.tools.bash import BashTool
@@ -18,74 +14,6 @@ from opencollab.adapters.tools.git_diff import GitDiffTool
 from opencollab.adapters.tools.human import AskUserTool
 from opencollab.adapters.tools.run_tests import RunTestsTool
 from opencollab.application.tool_execution import ToolRuntime
-
-
-def run(coro):
-    return asyncio.run(coro)
-
-
-class FakeEnv:
-    def __init__(self, stdout: str = ""):
-        self.stdout = stdout
-        self.exec_calls = []
-
-    async def exec_cmd(self, cmd: str, timeout: float = 120.0):
-        self.exec_calls.append((cmd, timeout))
-        return SimpleNamespace(returncode=0, stdout=self.stdout, stderr="")
-
-    async def read_file(self, path: str) -> str:
-        raise AssertionError("read_file was not expected")
-
-    async def write_file(self, path: str, content: str) -> None:
-        raise AssertionError("write_file was not expected")
-
-
-class SpySafetyPolicy:
-    def __init__(self):
-        self.cmd_calls = []
-        self.path_calls = []
-
-    def check_path(self, target_path: str) -> str:
-        self.path_calls.append(target_path)
-        return target_path
-
-    def check_cmd(self, cmd: str) -> None:
-        pass
-
-    def is_risky(self, cmd: str) -> bool:
-        return False
-
-    async def check_cmd_interactive(self, cmd: str, confirm_fn=None) -> None:
-        self.cmd_calls.append((cmd, confirm_fn))
-
-
-class FakeRemoteEnv:
-    """Environment that does I/O somewhere other than the host filesystem."""
-
-    def __init__(self, files: dict[str, str] | None = None):
-        self.files = dict(files or {})
-
-    async def read_file(self, path: str) -> str:
-        if path not in self.files:
-            raise FileNotFoundError(path)
-        return self.files[path]
-
-    async def write_file(self, path: str, content: str) -> None:
-        self.files[path] = content
-
-
-class SpyLock:
-    instances: list["SpyLock"] = []
-
-    def __init__(self, *args, **kwargs):
-        SpyLock.instances.append(self)
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *exc):
-        return False
-
 
 # ---------------------------------------------------------------------------
 # ToolRuntime wiring
@@ -186,565 +114,6 @@ def test_bash_tool_preserves_blocked_command_behavior(tmp_path):
 
     assert env.exec_calls == []
 
-
-# ---------------------------------------------------------------------------
-# FileReadTool
-# ---------------------------------------------------------------------------
-
-
-def test_file_read_reads_through_runtime(tmp_path):
-    workspace = tmp_path / "ws"
-    workspace.mkdir()
-    target = workspace / "note.txt"
-    target.write_text("alpha\nbeta\ngamma\n", encoding="utf-8")
-    env = LocalEnvironment(str(workspace))
-    sandbox = SandboxInterceptor(str(workspace))
-    runtime = ToolRuntime(environment=env, safety_policy=sandbox, permission_policy=None)
-
-    result = run(
-        FileReadTool().execute_with_runtime(
-            {"path": "note.txt", "offset": 2, "limit": 1},
-            runtime,
-        )
-    )
-
-    assert "File: note.txt (3 lines total, showing 2-2)" in result
-    assert "2\tbeta" in result
-    assert "1\talpha" not in result
-
-
-def test_file_read_small_range_bypasses_full_file_limit(tmp_path):
-    target = tmp_path / "large.txt"
-    target.write_bytes(b"first\nsecond\n" + b"x" * (5 * 1024 * 1024))
-    runtime = ToolRuntime(
-        environment=LocalEnvironment(str(tmp_path)),
-        safety_policy=None,
-        permission_policy=None,
-    )
-
-    result = run(
-        FileReadTool().execute_with_runtime(
-            {"path": "large.txt", "offset": 1, "limit": 2},
-            runtime,
-        )
-    )
-
-    assert "1\tfirst" in result
-    assert "2\tsecond" in result
-    assert "more lines below" in result
-
-
-def test_file_read_preserves_workspace_path_jail(tmp_path):
-    workspace = tmp_path / "ws"
-    workspace.mkdir()
-    env = LocalEnvironment(str(workspace))
-    sandbox = SandboxInterceptor(str(workspace))
-    runtime = ToolRuntime(environment=env, safety_policy=sandbox, permission_policy=None)
-
-    with pytest.raises(PermissionError):
-        run(FileReadTool().execute_with_runtime({"path": "/etc/passwd"}, runtime))
-
-
-def test_file_read_requires_environment():
-    runtime = ToolRuntime(environment=None, safety_policy=None, permission_policy=None)
-
-    result = run(FileReadTool().execute_with_runtime({"path": "missing.txt"}, runtime))
-
-    assert result == "Error: no execution environment available."
-
-
-def test_file_read_preserves_file_not_found(tmp_path):
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(FileReadTool().execute_with_runtime({"path": "missing.txt"}, runtime))
-
-    assert result == "Error: file not found: missing.txt"
-
-
-def test_file_read_preserves_permission_error_string(monkeypatch, tmp_path):
-    async def raise_permission_error(*args, **kwargs):
-        raise PermissionError("denied")
-
-    monkeypatch.setattr(LocalEnvironment, "read_text_range", raise_permission_error)
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(FileReadTool().execute_with_runtime({"path": "secret.txt"}, runtime))
-
-    assert result == "Error: denied"
-
-
-# ---------------------------------------------------------------------------
-# FileWriteTool
-# ---------------------------------------------------------------------------
-
-
-def test_file_write_create_and_str_replace(tmp_path):
-    workspace = tmp_path / "ws"
-    workspace.mkdir()
-    env = LocalEnvironment(str(workspace))
-    sandbox = SandboxInterceptor(str(workspace))
-    runtime = ToolRuntime(environment=env, safety_policy=sandbox, permission_policy=None)
-
-    create_result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "create", "content": "alpha\n"},
-            runtime,
-        )
-    )
-    replace_result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "str_replace", "old_str": "alpha", "new_str": "beta"},
-            runtime,
-        )
-    )
-
-    assert "Created/wrote" in create_result
-    assert "Replaced in" in replace_result
-    assert (workspace / "note.txt").read_text(encoding="utf-8") == "beta\n"
-
-
-@pytest.mark.parametrize("existing", [False, True])
-def test_file_write_create_requires_explicit_content(tmp_path, existing):
-    target = tmp_path / "note.txt"
-    if existing:
-        target.write_text("KEEP", encoding="utf-8")
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "create"},
-            runtime,
-        )
-    )
-
-    assert result == "Error: content is required for create mode."
-    assert target.exists() is existing
-    if existing:
-        assert target.read_text(encoding="utf-8") == "KEEP"
-
-
-def test_file_write_create_allows_explicit_empty_content(tmp_path):
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "empty.txt", "mode": "create", "content": ""},
-            runtime,
-        )
-    )
-
-    assert "Created/wrote" in result
-    assert (tmp_path / "empty.txt").read_text(encoding="utf-8") == ""
-
-
-def test_file_write_preserves_workspace_path_jail(tmp_path):
-    workspace = tmp_path / "ws"
-    workspace.mkdir()
-    env = LocalEnvironment(str(workspace))
-    sandbox = SandboxInterceptor(str(workspace))
-    runtime = ToolRuntime(environment=env, safety_policy=sandbox, permission_policy=None)
-
-    with pytest.raises(PermissionError):
-        run(
-            FileWriteTool().execute_with_runtime(
-                {"path": "/tmp/outside.txt", "mode": "create", "content": "nope"},
-                runtime,
-            )
-        )
-
-
-def test_file_write_requires_environment():
-    runtime = ToolRuntime(environment=None, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "create", "content": "x"},
-            runtime,
-        )
-    )
-
-    assert result == "Error: no execution environment available."
-
-
-def test_file_write_preserves_missing_old_str(tmp_path):
-    target = tmp_path / "note.txt"
-    target.write_text("alpha\n", encoding="utf-8")
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "str_replace", "new_str": "beta"},
-            runtime,
-        )
-    )
-
-    assert result == "Error: old_str is required for str_replace mode."
-
-
-def test_file_write_preserves_duplicate_old_str_error(tmp_path):
-    target = tmp_path / "note.txt"
-    target.write_text("alpha\nalpha\n", encoding="utf-8")
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "str_replace", "old_str": "alpha", "new_str": "beta"},
-            runtime,
-        )
-    )
-
-    assert (
-        result
-        == "Error: old_str found 2 times in note.txt. Provide more context to make it unique."
-    )
-
-
-def test_file_write_rejects_overlapping_old_str_matches(tmp_path):
-    target = tmp_path / "note.txt"
-    target.write_text("aaa", encoding="utf-8")
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "str_replace", "old_str": "aa", "new_str": "b"},
-            runtime,
-        )
-    )
-
-    assert (
-        result
-        == "Error: old_str found 2 times in note.txt. Provide more context to make it unique."
-    )
-    assert target.read_text(encoding="utf-8") == "aaa"
-
-
-# ---------------------------------------------------------------------------
-# Host write lock — only taken when I/O hits the host filesystem
-# ---------------------------------------------------------------------------
-
-
-def test_file_write_remote_env_takes_no_host_lock(monkeypatch):
-    SpyLock.instances = []
-    monkeypatch.setattr(base, "FileLock", SpyLock)
-    env = FakeRemoteEnv({"note.txt": "alpha\n"})
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "str_replace", "old_str": "alpha", "new_str": "beta"},
-            runtime,
-        )
-    )
-
-    assert "Replaced in" in result
-    assert env.files["note.txt"] == "beta\n"
-    assert SpyLock.instances == []
-
-
-def test_apply_patch_remote_env_takes_no_host_lock(monkeypatch):
-    SpyLock.instances = []
-    monkeypatch.setattr(base, "FileLock", SpyLock)
-    env = FakeRemoteEnv({"f.py": "a\nb\nc\nd\n"})
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        ApplyPatchTool().execute_with_runtime(
-            {
-                "path": "f.py",
-                "mode": "line_replace",
-                "start_line": 2,
-                "end_line": 2,
-                "new_str": "B",
-            },
-            runtime,
-        )
-    )
-
-    assert "Applied line_replace" in result
-    assert env.files["f.py"] == "a\nB\nc\nd\n"
-    assert SpyLock.instances == []
-
-
-def test_host_write_lock_files_live_outside_target_workspace(tmp_path):
-    env = LocalEnvironment(str(tmp_path))
-    lock_path = Path(base._host_lock_path("x.txt", env))
-
-    async def exercise():
-        async with base.host_write_lock("x.txt", env) as lock:
-            assert isinstance(lock, base.FileLock)
-            assert lock_path.exists()
-
-    run(exercise())
-
-    assert tmp_path not in lock_path.parents
-    assert list(tmp_path.rglob("*.lock")) == []
-
-
-class SlowLocalWriteEnv(FakeRemoteEnv):
-    local_filesystem = True
-    workspace = "/virtual/workspace"
-
-    async def read_file(self, path: str) -> str:
-        await asyncio.sleep(0.05)
-        return await super().read_file(path)
-
-    async def write_file(self, path: str, content: str) -> None:
-        await asyncio.sleep(0.05)
-        await super().write_file(path, content)
-
-
-def test_concurrent_host_writes_do_not_block_event_loop_on_lock():
-    env = SlowLocalWriteEnv({"note.txt": "initial"})
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-    tool = FileWriteTool()
-
-    async def exercise():
-        await asyncio.wait_for(
-            asyncio.gather(
-                tool.execute_with_runtime(
-                    {
-                        "path": "note.txt",
-                        "mode": "create",
-                        "content": "first",
-                        "overwrite": True,
-                    },
-                    runtime,
-                ),
-                tool.execute_with_runtime(
-                    {
-                        "path": "note.txt",
-                        "mode": "create",
-                        "content": "second",
-                        "overwrite": True,
-                    },
-                    runtime,
-                ),
-            ),
-            timeout=1.0,
-        )
-
-    started = time.monotonic()
-    run(exercise())
-
-    assert time.monotonic() - started < 1.0
-    assert env.files["note.txt"] in {"first", "second"}
-
-
-def test_host_write_lock_never_pollutes_git_diff(tmp_path):
-    subprocess.run(["git", "init", "-q"], cwd=tmp_path, check=True)
-    env = LocalEnvironment(str(tmp_path))
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        FileWriteTool().execute_with_runtime(
-            {"path": "note.txt", "mode": "create", "content": "hello\n"},
-            runtime,
-        )
-    )
-    status = subprocess.run(
-        ["git", "status", "--short"],
-        cwd=tmp_path,
-        check=True,
-        capture_output=True,
-        text=True,
-    ).stdout
-
-    assert result.startswith("Created/wrote")
-    assert ".lock" not in status
-    assert list(tmp_path.rglob("*.lock")) == []
-
-
-# ---------------------------------------------------------------------------
-# GrepTool
-# ---------------------------------------------------------------------------
-
-
-def test_grep_tool_requires_environment():
-    runtime = ToolRuntime(environment=None, safety_policy=None, permission_policy=None)
-
-    result = run(GrepTool().execute_with_runtime({"pattern": "needle"}, runtime))
-
-    assert result == "Error: no execution environment available."
-
-
-def test_grep_tool_routes_search_path_through_safety_policy():
-    env = FakeEnv(stdout="src/app.py:1:needle\n")
-    safety = SpySafetyPolicy()
-    runtime = ToolRuntime(environment=env, safety_policy=safety, permission_policy=None)
-
-    result = run(
-        GrepTool().execute_with_runtime(
-            {"pattern": "needle", "path": "../outside", "glob": "*.py", "max_results": 3},
-            runtime,
-        )
-    )
-
-    assert result == "src/app.py:1:needle"
-    assert safety.path_calls == ["../outside"]
-    assert len(env.exec_calls) == 1
-    cmd, timeout = env.exec_calls[0]
-    assert "needle" in cmd
-    assert "../outside" in cmd
-    assert timeout == 30
-
-
-def test_grep_tool_rejects_non_integer_max_results_before_exec():
-    env = FakeEnv(stdout="src/app.py:1:needle\n")
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        GrepTool().execute_with_runtime(
-            {"pattern": "needle", "max_results": "1; touch pwned"},
-            runtime,
-        )
-    )
-
-    assert result == "Error: max_results must be an integer."
-    assert env.exec_calls == []
-
-
-def test_grep_tool_applies_max_results_globally():
-    env = FakeEnv(
-        stdout="\n".join(f"src/file_{index}.py:1:needle" for index in range(6))
-    )
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(
-        GrepTool().execute_with_runtime(
-            {"pattern": "needle", "max_results": 2},
-            runtime,
-        )
-    )
-
-    assert result.splitlines() == [
-        "src/file_0.py:1:needle",
-        "src/file_1.py:1:needle",
-    ]
-
-
-def test_grep_tool_does_not_fallback_after_normal_rg_no_match():
-    class NoMatchEnv(FakeEnv):
-        async def exec_cmd(self, cmd: str, timeout: float = 120.0):
-            self.exec_calls.append((cmd, timeout))
-            return SimpleNamespace(returncode=1, stdout="", stderr="")
-
-    env = NoMatchEnv()
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(GrepTool().execute_with_runtime({"pattern": "absent"}, runtime))
-
-    assert result == "No matches found for pattern: absent"
-    assert len(env.exec_calls) == 1
-    assert "grep -r" not in env.exec_calls[0][0]
-
-
-@pytest.mark.parametrize(
-    ("returncode", "stderr", "expected"),
-    [
-        (2, "regex parse error", "regex parse error"),
-        (2, "path does not exist", "path does not exist"),
-        (-1, "Command timed out after 30s", "timed out"),
-    ],
-)
-def test_grep_tool_reports_backend_failures(returncode, stderr, expected):
-    class FailedSearchEnv(FakeEnv):
-        async def exec_cmd(self, cmd: str, timeout: float = 120.0):
-            self.exec_calls.append((cmd, timeout))
-            return SimpleNamespace(
-                returncode=returncode,
-                stdout="",
-                stderr=stderr,
-                stdout_truncated=False,
-                stderr_truncated=False,
-            )
-
-    env = FailedSearchEnv()
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(GrepTool().execute_with_runtime({"pattern": "["}, runtime))
-
-    assert result.startswith("Error: rg search failed")
-    assert expected in result
-    assert "No matches" not in result
-    assert "2>/dev/null" not in env.exec_calls[0][0]
-
-
-def test_grep_tool_includes_hidden_project_paths_with_noise_exclusions():
-    env = FakeEnv(stdout=".github/workflows/ci.yml:1:needle\n")
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    result = run(GrepTool().execute_with_runtime({"pattern": "needle"}, runtime))
-
-    assert result == ".github/workflows/ci.yml:1:needle"
-    command = env.exec_calls[0][0]
-    assert "rg -n --hidden " in command
-    assert "-g '!.git/**'" in command
-    assert "-g '!.venv/**'" in command
-    assert "-g '!.opencollab/**'" in command
-
-
-def test_grep_tool_terminates_options_before_pattern_and_path():
-    env = FakeEnv(stdout="")
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    run(
-        GrepTool().execute_with_runtime(
-            {"pattern": "--pre=touch pwned", "path": "--debug"},
-            runtime,
-        )
-    )
-
-    cmd, _ = env.exec_calls[0]
-    assert "rg -n --hidden" in cmd
-    assert "-- '--pre=touch pwned' --debug" in cmd
-
-
-def test_file_read_description_teaches_distill_and_forbids_reread():
-    # The file_read description is the UNIVERSAL carrier of the anti-thrash rule:
-    # every workflow/team that has the tool sees it, regardless of its prompt. Pins
-    # the rule (distill into notes; trust the cleared stub; don't re-read to
-    # reconfirm) motivated by a 90-read/0-write/step-cap stall.
-    desc = FileReadTool.description
-    assert "Distill as you read" in desc
-    assert "empty reply" in desc  # forbids bare tool-call turns
-    assert "re-read" in desc      # trust the stub instead
-
-
-def test_grep_tool_fallback_uses_ere_and_skips_session_dir():
-    # When rg is missing from PATH the command falls back to grep. That fallback
-    # MUST use -E (ERE) so `a|b|c` alternation works — plain `grep -rn` is BRE,
-    # where `|` is a literal char and an alternation pattern silently matches
-    # nothing. It must also skip .opencollab so it can't "match" its own logged
-    # pattern strings instead of real source. (Regression: a 100-step run stalled
-    # because every alternation grep returned "No matches found".)
-    class MissingRgEnv(FakeEnv):
-        async def exec_cmd(self, cmd: str, timeout: float = 120.0):
-            self.exec_calls.append((cmd, timeout))
-            if len(self.exec_calls) == 1:
-                return SimpleNamespace(returncode=127, stdout="", stderr="")
-            return SimpleNamespace(returncode=1, stdout="", stderr="")
-
-    env = MissingRgEnv(stdout="")
-    runtime = ToolRuntime(environment=env, safety_policy=None, permission_policy=None)
-
-    run(
-        GrepTool().execute_with_runtime(
-            {"pattern": "Scheduler|RowStatus"}, runtime
-        )
-    )
-
-    assert len(env.exec_calls) == 2
-    cmd, _ = env.exec_calls[1]
-    assert cmd.startswith("grep -rEn")
-    assert " grep -rn " not in cmd  # the buggy BRE fallback must be gone
-    assert "--exclude-dir=.opencollab" in cmd
 
 
 # ---------------------------------------------------------------------------
